+++ packages/ai-api/src/client/AI_CORE_API/schema/ai-scenario-label.ts
@@ -3,8 +3,9 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import type { AiScenarioLabelKey } from './ai-scenario-label-key.js';
+
 import type { AiLabelValue } from './ai-label-value.js';
+import type { AiScenarioLabelKey } from './ai-scenario-label-key.js';
 /**
  * Representation of the 'AiScenarioLabel' schema.
  */