+++ packages/ai-api/src/client/AI_CORE_API/schema/bcknd-argo-cd-application-deletion-response.ts
@@ -3,8 +3,9 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import type { BckndId } from './bcknd-id.js';
+
 import type { BckndArgoCDApplicationDeletionResponseMessage } from './bcknd-argo-cd-application-deletion-response-message.js';
+import type { BckndId } from './bcknd-id.js';
 /**
  * Representation of the 'BckndArgoCDApplicationDeletionResponse' schema.
  */