+++ packages/document-grounding/src/client/api/schema/search-filter.ts
@@ -3,8 +3,9 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import type { SearchConfiguration } from './search-configuration.js';
+
 import type { KeyValueListPair } from './key-value-list-pair.js';
+import type { SearchConfiguration } from './search-configuration.js';
 import type { SearchDocumentKeyValueListPair } from './search-document-key-value-list-pair.js';
 /**
  * Representation of the 'SearchFilter' schema.