+++ packages/foundation-models/src/azure-openai/client/inference/schema/azure-search-chat-extension-parameters.ts
@@ -3,13 +3,14 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import type { AzureOpenAiOnYourDataApiKeyAuthenticationOptions } from './on-your-data-api-key-authentication-options.js';
-import type { AzureOpenAiOnYourDataSystemAssignedManagedIdentityAuthenticationOptions } from './on-your-data-system-assigned-managed-identity-authentication-options.js';
-import type { AzureOpenAiOnYourDataUserAssignedManagedIdentityAuthenticationOptions } from './on-your-data-user-assigned-managed-identity-authentication-options.js';
+
 import type { AzureOpenAiAzureSearchIndexFieldMappingOptions } from './azure-search-index-field-mapping-options.js';
 import type { AzureOpenAiAzureSearchQueryType } from './azure-search-query-type.js';
-import type { AzureOpenAiOnYourDataEndpointVectorizationSource } from './on-your-data-endpoint-vectorization-source.js';
+import type { AzureOpenAiOnYourDataApiKeyAuthenticationOptions } from './on-your-data-api-key-authentication-options.js';
 import type { AzureOpenAiOnYourDataDeploymentNameVectorizationSource } from './on-your-data-deployment-name-vectorization-source.js';
+import type { AzureOpenAiOnYourDataEndpointVectorizationSource } from './on-your-data-endpoint-vectorization-source.js';
+import type { AzureOpenAiOnYourDataSystemAssignedManagedIdentityAuthenticationOptions } from './on-your-data-system-assigned-managed-identity-authentication-options.js';
+import type { AzureOpenAiOnYourDataUserAssignedManagedIdentityAuthenticationOptions } from './on-your-data-user-assigned-managed-identity-authentication-options.js';
 /**
  * Parameters for Azure Search when used as an Azure OpenAI chat extension.
  */