+++ packages/rpt/src/client/rpt/schema/predict-response-payload.ts
@@ -3,9 +3,10 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
+
+import type { PredictResponseMetadata } from './predict-response-metadata.js';
 import type { PredictResponseStatus } from './predict-response-status.js';
 import type { PredictionResult } from './prediction-result.js';
-import type { PredictResponseMetadata } from './predict-response-metadata.js';
 /**
  * Response payload for prediction requests.
  * Contains a list of prediction results.
@@ -15,6 +16,9 @@ export type PredictResponsePayload = {
    * Unique ID for the request.
    */
   id: string;
+  /**
+   * Status message that can indicate warnings (e.g. about suboptimal data).
+   */
   status: PredictResponseStatus;
   /**
    * Mapping of column names to their list of prediction results or index column.