+++ packages/foundation-models/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ai-sdk/foundation-models
 
+## 2.15.0
+
+### Patch Changes
+
+- Updated dependencies [a5407c8]
+  - @sap-ai-sdk/core@2.15.0
+  - @sap-ai-sdk/ai-api@2.15.0
+
+## 2.14.0
+
+### Patch Changes
+
+- Updated dependencies [cca7edd]
+- Updated dependencies [cca7edd]
+  - @sap-ai-sdk/core@2.14.0
+  - @sap-ai-sdk/ai-api@2.14.0
+
 ## 2.13.0
 
 ### Patch Changes